change location of test
