+++ tests/ludwig/utils/automl/test_type_inference.py
@@ -1,10 +1,7 @@
 import random
 
-import numpy as np
-import pandas as pd
 import pytest
 
-from ludwig.automl.base_config import get_dataset_info
 from ludwig.constants import AUDIO, BINARY, CATEGORY, DATE, IMAGE, NUMBER, TEXT
 from ludwig.data.dataset_synthesizer import generate_string
 from ludwig.utils.automl.field_info import FieldInfo
@@ -94,18 +91,3 @@ def test_auto_type_inference_single_value_binary_feature():
     )
     assert infer_type(field=field, missing_value_percent=0, row_count=ROW_COUNT) == CATEGORY
     assert should_exclude(idx=3, field=field, dtype="object", row_count=ROW_COUNT, targets={TARGET_NAME})
-
-
-@pytest.mark.parametrize("missing_value", [(None), (np.nan)], ids=["none", "np.nan"])
-@pytest.mark.distributed
-def test_bool_type_inference_from_object_dtype(missing_value):
-    df = pd.DataFrame({"col1": ["a", "b", "a", "a", "b", missing_value]})
-    info = get_dataset_info(df)
-    assert info.fields[0].dtype == "bool"
-
-
-@pytest.mark.distributed
-def test_object_type_inference_from_object_dtype():
-    df = pd.DataFrame({"col1": ["a", "b", "c", "d", "e", "a", "b", "b"]})
-    info = get_dataset_info(df)
-    assert info.fields[0].dtype == "object"
